Update script.sh
In a nutshell, I made several fixes and improvements to the script to make it work on most Linux distributions. I used more flexible and adaptable commands and conditions, and also adjusted the directories and paths to be compatible with different systems. This will make the script more versatile and able to run smoothly on a wide variety of Linux configurations.

For more details contact me: @KirbyFrog